# api/openapi.yaml
openapi: 3.1.0
info:
  title: Personal Intelligence and Action Engine API
  version: 0.1.0
  description:
    Initial contract for authenticated workspaces, knowledge, conversations, memory, tools, approvals, audit, and
    evaluations.
servers:
  - url: /
security:
  - bearerAuth: []
...
paths:
...
  /v1/workspaces/{workspace_id}/retrieval/query:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
    post:
      tags:
        - Retrieval
      operationId: queryKnowledge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RetrievalQuery'
      responses:
        '200':
          description: Authorized retrieval results.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetrievalResponse'
  /v1/workspaces/{workspace_id}/retrieval/traces/{trace_id}:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
      - name: trace_id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
    get:
      tags:
        - Retrieval
      operationId: getRetrievalTrace
      responses:
        '200':
          description: Retrieval trace visible to authorized role.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetrievalResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    WorkspaceId:
      name: workspace_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Uuid'
...
  responses:
...
    NotFound:
      description: Resource not found or not visible to this principal.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
    Uuid:
      type: string
      format: uuid
...
    ErrorEnvelope:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
            - request_id
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              $ref: '#/components/schemas/Uuid'
            details:
              type: object
              additionalProperties: true
...
    RetrievalQuery:
      type: object
      additionalProperties: false
      required:
        - query
      properties:
        query:
          type: string
          minLength: 1
          maxLength: 10000
        project_id:
          oneOf:
            - $ref: '#/components/schemas/Uuid'
            - type: 'null'
        source_ids:
          type: array
          maxItems: 100
          items:
            $ref: '#/components/schemas/Uuid'
        history_mode:
          type: string
          enum:
            - CURRENT_ONLY
            - INCLUDE_HISTORY
          default: CURRENT_ONLY
        limit:
          type: integer
          minimum: 1
          maximum: 50
          default: 10
        include_debug:
          type: boolean
          default: false
    RetrievalResult:
      type: object
      required:
        - rank
        - chunk_id
        - document_id
        - document_version_id
        - source_id
        - locator
        - text
        - scores
      properties:
        rank:
          type: integer
          minimum: 1
        chunk_id:
          $ref: '#/components/schemas/Uuid'
        document_id:
          $ref: '#/components/schemas/Uuid'
        document_version_id:
          $ref: '#/components/schemas/Uuid'
        source_id:
          nullable: true
          $ref: '#/components/schemas/Uuid'
        source_title:
          type: string
        locator:
          $ref: '#/components/schemas/SourceLocator'
        text:
          type: string
        scores:
          type: object
          required:
            - fused
          properties:
            lexical:
              type:
                - number
                - 'null'
            vector:
              type:
                - number
                - 'null'
            fused:
              type: number
    RetrievalResponse:
      type: object
      required:
        - trace_id
        - configuration_version
        - results
        - latency_ms
      properties:
        trace_id:
          $ref: '#/components/schemas/Uuid'
        configuration_version:
          type: string
        results:
          type: array
          items:
            $ref: '#/components/schemas/RetrievalResult'
        latency_ms:
          type: integer
          minimum: 0
    SourceLocator:
      type: object
      additionalProperties: true
      properties:
        page:
          type:
            - integer
            - 'null'
        section:
          type:
            - string
            - 'null'
        paragraph:
          type:
            - integer
            - 'null'
        start_char:
          type:
            - integer
            - 'null'
        end_char:
          type:
            - integer
            - 'null'
...